pyproject.toml
[project]
name = "tenko"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "beautifulsoup4>=4.12.0",
    "brotli>=1.2.0",
    "bs4>=0.0.2",
    "certifi>=2025.11.12",
    "charset-normalizer>=3.4.4",
    "idna>=3.11",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=21.0",
    "requests>=2.32.5",
    "urllib3>=2.5.0",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.17.0",
]
